# docs/contracts/0001-service-api-user-auth.openapi.yaml
openapi: 3.1.1
info:
  title: quant-v2 service-api
  version: 1.0.0
  description: Versioned User and authentication API. All business routes use /api/v1.
servers:
  - url: http://127.0.0.1:13000
paths:
  /api/v1/auth/login:
    post:
      tags: [auth]
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Access token issued; HttpOnly refresh_token cookie set.
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessToken'
        '401':
          $ref: '#/components/responses/Problem'
        '400':
          $ref: '#/components/responses/Problem'
        '429':
          $ref: '#/components/responses/Problem'
        '503':
          $ref: '#/components/responses/Problem'
  /api/v1/auth/refresh:
    post:
      tags: [auth]
      operationId: refresh
      parameters:
        - $ref: '#/components/parameters/RefreshCookie'
      responses:
        '200':
          description: Refresh token rotated and access token issued.
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessToken'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '429':
          $ref: '#/components/responses/Problem'
        '503':
          $ref: '#/components/responses/Problem'
  /api/v1/auth/logout:
    post:
      tags: [auth]
      operationId: logout
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Current session revoked and refresh cookie cleared.
        '401':
          $ref: '#/components/responses/Problem'
  /api/v1/users/me:
    get:
      tags: [users]
      operationId: getCurrentUser
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Problem'
    patch:
      tags: [users]
      operationId: updateCurrentUser
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Updated user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
  /api/v1/users/me/password:
    post:
      tags: [users]
      operationId: changeCurrentUserPassword
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '204':
          description: Password changed; old sessions become invalid through securityVersion.
        '401':
          $ref: '#/components/responses/Problem'
  /api/v1/users:
    get:
      tags: [users]
      operationId: listUsers
      security:
        - bearerAuth: []
      parameters:
        - name: cursor
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Cursor page of users.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPage'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
    post:
      tags: [users]
      operationId: createUser
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
  /api/v1/users/{id}:
    patch:
      tags: [users]
      operationId: updateUser
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    UserId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    RefreshCookie:
      name: refresh_token
      in: cookie
      required: true
      schema:
        type: string
  responses:
    Problem:
      description: RFC 9457 problem response.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
    LoginRequest:
      type: object
      required: [email, password]
      additionalProperties: false
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AccessToken:
      type: object
      required: [accessToken]
      properties:
        accessToken:
          type: string
    User:
      type: object
      required: [id, email, displayName, role, status, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        displayName:
          type: string
        role:
          type: string
          enum: [USER, ADMIN]
        status:
          type: string
          enum: [ACTIVE, DISABLED]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UserPage:
      type: object
      required: [items, page]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/User'
        page:
          type: object
          required: [nextCursor]
          properties:
            nextCursor:
              type: [string, 'null']
              format: uuid
    UpdateProfileRequest:
      type: object
      required: [displayName]
      additionalProperties: false
      properties:
        displayName:
          type: string
          minLength: 1
          maxLength: 120
    ChangePasswordRequest:
      type: object
      required: [currentPassword, newPassword]
      additionalProperties: false
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
          minLength: 12
          pattern: '\\d'
    CreateUserRequest:
      type: object
      required: [email, password, displayName]
      additionalProperties: false
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 12
          pattern: '\\d'
        displayName:
          type: string
          minLength: 1
          maxLength: 120
        role:
          type: string
          enum: [USER, ADMIN]
        status:
          type: string
          enum: [ACTIVE, DISABLED]
    UpdateUserRequest:
      type: object
      additionalProperties: false
      properties:
        displayName:
          type: string
          minLength: 1
          maxLength: 120
        role:
          type: string
          enum: [USER, ADMIN]
        status:
          type: string
          enum: [ACTIVE, DISABLED]
    Problem:
      type: object
      required: [type, title, status, detail, instance]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        requestId:
          type: string